Basic PHPLint support
Move src to a directory named after the namespace RemexHtml, to support
PHPLint's idea of autoloading.

Ran PHPLint 3.0 and fixed a few minor issues. There are still many, many
errors, and it's not a goal to fix them. But I was able to verify that there
are no unintentional missing breaks, or variables accessed before their
assignment.

Change-Id: Icb269d934a80bdfe0332c0003c854bffdc58eb71