Serializer::endTag() removes descendants from the node list
Reduce memory usage by removing descendant nodes when we get an endTag()
event. Usually an endTag() event would be received for the children
before the parent, but in some cases, the endTag() event is omitted, so
it makes sense to clean up, since we are traversing the descendant tree
anyway. Mutation on descendant nodes after endTag() is certainly
invalid, as demonstrated by T200827.

Also add some tracing of live nodes when endDocument() is received.

Also rename stringify to serializeNode.

Bug: T200827
Change-Id: Ia232ad17738efc8a2f4c3dc16e11c984b71beb55