Don't double-decode HTML entities on non-HHVM PHP
When given the string '&amp;amp;' the Tokenizer::handleCharRefs() function
will attempt to "shortcut" the decode using html_entity_decode() when
not running on HHVM, and *then* will proceed to decode the entities again
the "long way".  This leads to an erroneous double-decode.

Just remove the entire shortcut.

This shows up in a double-decode of (eg) `[[File:Foobar.jpg|alt=&amp;amp;]]`
in MediaWiki.

Bug: T207088
Change-Id: Ide67bba20f771868c0e119cb2874464dcf1d758a