Optimize parsing of "simple tags" (no attributes, no funny characters)
This is an optimization ported from domino's HTML parser: if the
tag has no attributes don't bother switching to attribute parsing,
just stay in the data state and immediately emit the tag.

This improves the parse-to-DOM time of [[en:Barack Obama]] on my machine
from 6.8s to 6.3s (there are still a lot of attributes in that test case,
and parsing them is slow).

Change-Id: Ic4a8e155b7ca35fac62f3ee3ddff8875a93873de