Efficiently handle well-behaved character references
Attribute values typically only encode only &amp; and either &apos; or &quot,
depending on which is used to quote the string.  Sometimes you see &lt; and
&gt; quoted, even though the HTML5 doesn't require this.  Efficiently handle
attribute values which only contain these entities (and where all entities
are semicolon-terminated), only falling back to the full spec-compliant
entity decoder if we find text which doesn't match the above.

This improves the performance of parsing [[en:Barack Obama]] from 4.75s
to 4.57s on my machine.  Passing `ignoreCharRefs` and `ignoreNulls` to
the tokenizer construct yields a 4.43s parse, so the penalty for
a spec-compliant parse is getting very small.  (And we haven't yet
optimized decode of the Text nodes at all.)

Change-Id: I52c567a46e15e6db97840d2b7e3eeaf6acac41ec