Optimize handling of "simple" data regions
When searching for tags in the data state, use strcspn to skip ahead
to the next "difficult" character before starting the regular expression
match.  If it succeeds at no offset, when we know that the region skipped
doesn't have any nulls, <, or character entities in it, and we can skip
those processing steps.  This actually improves performance even if
`ignoreCharRefs`/`ignoreNulls`/`ignoreErrors` is set, because `strcspn`
is faster than `preg_match` at the basic string-scanning task.

Before this patch, parsing [[en:Barak Obama]] to DOM with only
`suppressHtmlNamespace` set takes 4.59s; turning on `ignoreCharRefs`
and `ignoreNulls` yields 4.43s.

After this patch, parsing to DOM with only `suppressHtmlNamespace` set
takes 4.38s.  Turning on `ignoreCharRefs` and `ignoreNulls` yields a
time of 4.38s (identical to w/in measurement error).

Turning on `ignoreErrors` drops the time down further to 4.17s (regardless
of whether `ignoreCharRefs`/`ignoreNulls` is on or off).

Change-Id: I4712060902f20d1dafb4b1153604342abde5fa35