Avoid 'u' flag in HTML Tokenizer preprocess step
Using the pcre 'u' flag turns off the pcre JIT, negatively impacting
performance.  Rewrite to use alternations instead of character classes
so that we can run the preprocessing regexp on the UTF-8 bytestream
directly.

Also use static $re to ensure that the same string object is passed
to pcre every time (although it should be an interned compile-time
constant anyway), and use strcspn to skip the \r replacement where
it is not necessary.

Change-Id: Ic27f0afb85e055a4d10f06f120de91e7f22e5985