Use compile-time constant string for character reference regexp
The PCRE regexp cache uses the string as the hashtable key.
If the string is not a compile-time constant, a complete string
comparison is required in order to validate the hash match.  The
character reference regexp is ~26kB long, so the string comparison
takes a substantial chunk of time.

Rewrite the regexp as a compile-time constant in the HTMLData class
to ensure it is able to be looked up quickly in the regexp cache.

See https://marc.info/?l=php-internals&m=155371149031157&w=2 for
further discussion of the performance issue in PHP and various
strategies to mitigate.

Change-Id: If98cdfc01d7fdfc802b1784fc9ce586d290238fd