Add scriptingFlag option to Tokenizer, and make it true by default
Previously the scriptingFlag was undeclared and implicitly false. Make
it true for consistency with the same flag in TreeBuilder. This has the
sole effect of causing <noscript> tags to have the "raw text" content
model.

Also other minor fixes and documentation changes suggested by PHPStorm
analysis. Remove unused variables etc.

Add ext-dom to the composer.json as a dependency -- I avoided doing this
in the past since you can use the non-DOM pieces without the DOM
extension, but maybe it's better to err on the side of too many
dependencies than too few. Better for composer to complain unnecessarily
than to have a fatal error at runtime.

Change-Id: I89371cc7088648651c88f312a63c0abc688463b6