Call the non-standard \DOMElement::setIdAttribute() method by default
This ensures that id attributes are actually indexed, as a modern HTML
DOM expects, and thus that DOMDocument::getElementById() is O(1) instead
of O(N).

Changed `DOMBuilder::createDocument` to protected, from private, so that
a subclass could be written to use a different DOMImplementation to work
around this and other PHP DOM "quirks".  In that case you'd set the new
`suppressIdAttribute` option to `true` in order to ensure that only
standard DOM methods are called.

Bug: T232390
Change-Id: Ib38b9bf2f1c799344a1b4a255e781e681961279d