Allow constructed DOMImplementation object to be passed to DOMBuilder
The DOM spec doesn't actually say that DOMImplementation objects will
have constructors.  In fact, in DOM Level 3 you're supposed to get a
DOMImplementation object by calling the `getDOMImplementation` method
on a DOMImplementationSource... which just kicks the can up the
stairs: where do you get a DOMImplementationSource object?

In any case, assuming you've managed to get your hands on a
DOMImplementation object, allow the user to pass it into DOMBuilder,
instead of asking Remex to call a constructor which does not appear in
any spec.

Followup-To: I5cbfa2374af37a361e00f79a1a4f2f13bf8a631b
Change-Id: I47b9540389fd744486879b7dbb0f12dd5b844cfe