Modernize 2nd argument to DOMImplementation::createDocument()
The 2nd argument to DOMImplementation::createDocument() is marked
DOMString with the [[LegacyNullToEmptyString]] attribute, which means
nulls are technically allowed but are expected to be converted to
the empty string by the implementation.  Avoid relying on this legacy
behavior and just pass an empty string as the second argument instead
of null.

Change-Id: I512bf34d7093bea7ec5d96644a964e76472fb9ba