Fix handling of <body> tag in "after head" state
We were missing the case for <body> in the "after head" state, which
would cause the tree builder to create a new synthetic <body> tag, and
then process the <body> tag that it saw in the "in body" state.
Luckily, when in the "in body" state seeing a <body> tag will result
in the attributes of the "second" body being merged with the
already-existing body tag, so the DOM which was built was correct but
(a) an error was flagged via:
  `$builder->error( 'merged unexpected body tag', $sourceStart );`
and (b) the $sourceLength of the built <body> tag was incorrectly set
to zero when the tag was created, reflecting that it was built as a
synthetic fixup tag.

See:
https://html.spec.whatwg.org/multipage/parsing.html#the-after-head-insertion-mode
https://html.spec.whatwg.org/multipage/parsing.html#parsing-main-inbody

Change-Id: Ib18efc7192d3b10c8b307da1a2f6a2e96a4f4b08