Change namespace from RemexHtml to Wikimedia\RemexHtml
This matches the recommendation in PSR-4 that the first component of
the namespace is the "vendor namespace" and makes the namespace match
the composer package name.

Aliases are set up so that the old Remex namespace will continue to
work.  In the next major release it is expected that the aliases
will be removed.

Change-Id: Icbe491222be86ca68236ad5ea42f0da31e7749a3