Add Dispatcher::flushTableText()
InTableText is a very special mode. It queues pending characters to
avoid fostering whitespace.

The old Parsoid shadow comment hack inserted a comment to avoid
fostering, which also exited the InTableText mode. The new idea is to
insert a meta tag more or less directly, but the one obstacle to doing
this is the lack of a way to flush the pending characters before meta
insertion.

So, provide a public method Dispatcher::flushTableText().

Change-Id: I1a5f1998547ece1e3bc9bda72f65d1b7c5fcfae7