Dodo compatibility for Element::$prefix and HtmlFormatter
The DOM spec says that prefix is a "nullable non-empty string"; the
actual PHP DOM implementation returns an empty string instead of null.
Use falsey instead of explicit tests against null or '' to keep the
code compatible with both the PHP DOM library and a
standards-compliant DOM library (like Dodo).

For similar reasons, remove some explicit type checks in the method
signatures of HtmlFormatter::formatNode() and ::formatDOMElement() to
make sure they don't break if an alternate DOM library is used.

Change-Id: I676097cdc65984968a3943d2450048e2c8e1ddfc