Make DOMBuilder::insertNode() protected, rather than private
The HTML5 specs has special rules for <template> elements: child nodes
are added not directly underneath the <template> but as children of
the DocumentFragment attached to the <template>.  Remex doesn't know
about this detail directly, but a spec-compliant DOM builder will want
to handle this in its ::insertNode() implementation.  In order to do
this, we make ::insertNode() protected (rather than private) and ensure
that text node insertion goes via ::insertNode() as well so there's
a single location to handle <template> nodes.

Change-Id: Ie9d96cc25e3f501a6cbc7c66c1edde9a6cbea3ce