Reduce memory usage of Tokenizer::consumeAttribs()
This shows up as a memory hotspot during profiling, and the "lazy
attributes" mechanism is in many cases unused: when parsing to DOM or
in a parse->modify->serialize pipeline every attribute will eventually
be demanded.

Preserve the "lazy attributes" functionality behind a tokenizer options
flag which defaults to off.  In cases where it is useful (for example,
stripping tags from input) this flag can be provided.

Change-Id: I34e8fd2b8a20a94b5ee42752bc2ad4286b73d0b7