Improve PHP 8.4 compatibility fixes
Clean up the PHP 8.4 compatibility fixes.

As described in T393922#10837089, we have three Remex goals with
respect to the new standards-compliant DOM implementation in PHP 8.4:

1. Isolate/remove the complex PHP < 8.4 \DOMDocument workarounds when
using PHP 8.4.

2. Try to map PHP 8.4 Dom\HTMLDocument::loadFromString() options to Remex
options so that clients can use whichever is faster for their particular
use case.  (This patch doesn't do much of this.)

3. Encapsulate the "coerced character set" issues documented in
documented in https://github.com/whatwg/dom/issues/849 and
https://github.com/whatwg/dom/issues/769. At the moment the DOM spec
has a narrower definition of "legal XML name" than the parser permits
-- and actual browsers in practice (and libxml2, used by PHP 8.4) use
an *even narrower* definition.  We expect eventually the DOM spec
(and PHP implementation) will accept a broader set of name characters,
but in the meantime we have three different workarounds we can apply.

Performance is identical to the previous patch, according to our
benchmark -- which is not surprising since all of the changes are
on corner cases which are not common in practice.

Bug: T393922
Change-Id: I9b7a98e9a8574834555534cca7a79b63cc2bf438