Further PHP 8.4 \Dom\Document fixes for DOMBuilder
Ensure that `<mw:section>` and `<math>` are both parsed correctly
with PHP 8.4, and that `suppressHtmlNamespace` works consistently,
*but only for tags in the HTML namespace*.

It is strongly recommended that `suppressHtmlNamespace` be set
to `true` when using the PHP < 8.4 \DOMImplementation, and to
`false` when using the PHP >= 8.4 \Dom\Implementation.

Note that there is a bug in \DOMImplementation when using ::importNode
on a tag with a non-null namespace: the `prefix` of the node can
be arbitrarily set to the string `default`.  That is not addressed
by this patch: `suppressHtmlNamespace` only affects nodes in the
HTML namespace, while this bug affects tags like <math> in non-HTML
namespaces.  It is recommended to either avoid `importNode` or
ignore the `prefix` on Nodes when using \DOMImplementation.

Bug: T400060
Follows-Up: I3502320cd3be21fcdba220b08618afd47f3f4422
Follows-Up: I9bc66ed5a604f7659503a851cc5ee53e7f4b3dfd
Change-Id: I989259414c2387629479b2947dd864233b29ec68